mcp-agent/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
paramiko==3.5.0
pyyaml==6.0.2
pydantic==2.10.0